Babel Plugin Async Load: Lazy load react components when necessary
