+++ components/code-editor/index.js
@@ -2,94 +2,105 @@
  * WordPress dependencies
  */
 import { Component } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
+import { UP, DOWN } from '@wordpress/keycodes';
 
-/**
- * Internal dependencies
- */
-import CodeEditor from './editor';
-import Placeholder from '../../packages/components/src/placeholder';
-import Spinner from '../../packages/components/src/spinner';
-
-function loadScript() {
-	return new Promise( ( resolve, reject ) => {
-		const handles = [ 'wp-codemirror', 'code-editor', 'htmlhint', 'csslint', 'jshint' ];
-
-		// Don't load htmlhint-kses unless we need it
-		if ( window._wpGutenbergCodeEditorSettings.htmlhint.kses ) {
-			handles.push( 'htmlhint-kses' );
-		}
-
-		const script = document.createElement( 'script' );
-		script.src = `${ wpApiSettings.schema.url }/wp-admin/load-scripts.php?load=${ handles.join( ',' ) }`;
-		script.onload = resolve;
-		script.onerror = reject;
-
-		document.head.appendChild( script );
-	} );
-}
-
-function loadStyle() {
-	return new Promise( ( resolve, reject ) => {
-		const handles = [ 'wp-codemirror', 'code-editor' ];
-
-		const style = document.createElement( 'link' );
-		style.rel = 'stylesheet';
-		style.href = `${ wpApiSettings.schema.url }/wp-admin/load-styles.php?load=${ handles.join( ',' ) }`;
-		style.onload = resolve;
-		style.onerror = reject;
-
-		document.head.appendChild( style );
-	} );
-}
-
-let hasAlreadyLoadedAssets = false;
-
-function loadAssets() {
-	if ( hasAlreadyLoadedAssets ) {
-		return Promise.resolve();
-	}
-
-	return Promise.all( [ loadScript(), loadStyle() ] ).then( () => {
-		hasAlreadyLoadedAssets = true;
-	} );
-}
-
-class LazyCodeEditor extends Component {
+class CodeEditor extends Component {
 	constructor() {
 		super( ...arguments );
 
-		this.state = {
-			status: 'pending',
-		};
+		this.onFocus = this.onFocus.bind( this );
+		this.onBlur = this.onBlur.bind( this );
+		this.onCursorActivity = this.onCursorActivity.bind( this );
+		this.onKeyHandled = this.onKeyHandled.bind( this );
 	}
 
 	componentDidMount() {
-		loadAssets().then(
-			() => {
-				this.setState( { status: 'success' } );
-			},
-			() => {
-				this.setState( { status: 'error' } );
+		const settings = this.props.settings || window._wpGutenbergCodeEditorSettings;
+		const instance = wp.codeEditor.initialize( this.textarea, settings );
+		this.editor = instance.codemirror;
+
+		this.editor.on( 'focus', this.onFocus );
+		this.editor.on( 'blur', this.onBlur );
+		this.editor.on( 'cursorActivity', this.onCursorActivity );
+		this.editor.on( 'keyHandled', this.onKeyHandled );
+
+		// Pass a reference to the editor back up.
+		if ( this.props.editorRef ) {
+			this.props.editorRef( this.editor );
+		}
+
+		this.updateFocus();
+	}
+
+	componentDidUpdate( prevProps ) {
+		if ( this.props.value !== prevProps.value && this.editor.getValue() !== this.props.value ) {
+			this.editor.setValue( this.props.value );
+		}
+
+		if ( this.props.focus !== prevProps.focus ) {
+			this.updateFocus();
+		}
+	}
+
+	componentWillUnmount() {
+		this.editor.on( 'focus', this.onFocus );
+		this.editor.off( 'blur', this.onBlur );
+		this.editor.off( 'cursorActivity', this.onCursorActivity );
+		this.editor.off( 'keyHandled', this.onKeyHandled );
+
+		this.editor.toTextArea();
+		this.editor = null;
+	}
+
+	onFocus() {
+		if ( this.props.onFocus ) {
+			this.props.onFocus();
+		}
+	}
+
+	onBlur( editor ) {
+		if ( this.props.onChange ) {
+			this.props.onChange( editor.getValue() );
+		}
+	}
+
+	onCursorActivity( editor ) {
+		this.lastCursor = editor.getCursor();
+	}
+
+	onKeyHandled( editor, name, event ) {
+		/*
+		 * Pressing UP/DOWN should only move focus to another block if the cursor is
+		 * at the start or end of the editor.
+		 *
+		 * We do this by stopping UP/DOWN from propagating if:
+		 *  - We know what the cursor was before this event; AND
+		 *  - This event caused the cursor to move
+		 */
+		if ( event.keyCode === UP || event.keyCode === DOWN ) {
+			const areCursorsEqual = ( a, b ) => a.line === b.line && a.ch === b.ch;
+			if ( this.lastCursor && ! areCursorsEqual( editor.getCursor(), this.lastCursor ) ) {
+				event.stopImmediatePropagation();
 			}
-		);
+		}
+	}
+
+	updateFocus() {
+		if ( this.props.focus && ! this.editor.hasFocus() ) {
+			// Need to wait for the next frame to be painted before we can focus the editor
+			window.requestAnimationFrame( () => {
+				this.editor.focus();
+			} );
+		}
+
+		if ( ! this.props.focus && this.editor.hasFocus() ) {
+			document.activeElement.blur();
+		}
 	}
 
 	render() {
-		if ( this.state.status === 'pending' ) {
-			return (
-				<Placeholder>
-					<Spinner />
-				</Placeholder>
-			);
-		}
-
-		if ( this.state.status === 'error' ) {
-			return <Placeholder>{ __( 'An unknown error occurred.' ) }</Placeholder>;
-		}
-
-		return <CodeEditor { ...this.props } />;
+		return <textarea ref={ ( ref ) => ( this.textarea = ref ) } defaultValue={ this.props.value } />;
 	}
 }
 
-export default LazyCodeEditor;
+export default CodeEditor;
