+++ packages/babel-plugin-async-load/package.json
@@ -1,0 +1,36 @@
+{
+	"name": "@wordpress/babel-plugin-async-load",
+	"version": "1.0.0",
+	"description": "Babel transform plugin for lazy load React components that depend on WordPress scripts and styles.",
+	"author": "The WordPress Contributors",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"babel-plugin",
+		"react"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/babel-plugin-async-load/README.md",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/gutenberg.git"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"files": [
+		"build",
+		"build-module"
+	],
+	"main": "build/index.js",
+	"module": "build-module/index.js",
+	"devDependencies": {
+		"@babel/core": "^7.0.0-beta.52",
+		"@babel/plugin-syntax-jsx": "^7.0.0-beta.52"
+	},
+	"peerDependencies": {
+		"@babel/core": "^7.0.0-beta.52"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
